// Front-end/OfferteApp/src/interfaces/Options.ts
export interface Option {
    id: number;
    name: string;
    spatrand: string;
    vensterbank: string;
    boorgatenPerStuk: boolean;
    wcd: boolean;
    randafwerking: boolean;
    prijsPerM2: number;
    randafwerkingPerM: number;
    spatrandPerM: number;
    vensterbankPerM: number;
    uitsparingOnderbouw: number;
    uitsparingInleg: number;
    uitsparingRuw: number;
    kraangat: number;
    zeepdispenser: number;
    boorgatenPerStukPrijs: number;
    wcdPrijs: number;
    achterwandPerM: number;
    randafwerkingPerMPrice: number;
}